requirements.txt
django
djangorestframework
factory_boy
